SHRINKWRAP-258 Moved StrictFilter to api-maven
* Moved StrictFilter to api-maven from impl-maven
* Improved MavenDependency JavaDoc
* Added one Exclusion test
