+++ extension-resolver/impl-maven/src/test/java/org/jboss/shrinkwrap/resolver/impl/maven/MavenResolutionFilterUnitTestCase.java
@@ -35,7 +35,7 @@ import org.jboss.shrinkwrap.resolver.api.ResolutionException;
 import org.jboss.shrinkwrap.resolver.api.maven.MavenDependencyResolver;
 import org.jboss.shrinkwrap.resolver.api.maven.filter.CombinedFilter;
 import org.jboss.shrinkwrap.resolver.api.maven.filter.ScopeFilter;
-import org.jboss.shrinkwrap.resolver.impl.maven.filter.StrictFilter;
+import org.jboss.shrinkwrap.resolver.api.maven.filter.StrictFilter;
 import org.junit.BeforeClass;
 import org.junit.Test;
